Fixed bug where characters did not properly leash
