+++ Assets/Resources/scripts/characters/ai/CommandStackScript.cs
@@ -2,15 +2,32 @@
 using System.Collections;
 using System.Collections.Generic;
 
-/* CommandStackScript - A Stack of commands that allow an enemy to figure out what he/she/it should be doing at any given time */
+/* CommandStackScript - A Stack of commands that allow an enemy to figure out what he/she/it should be doing at any given time 
+	* Commands are added to the stack if a given condition is met (i.e. Player enters aggro radius)
+	* Commands are removed from the stack if they meet their goal (i.e. 'walk to this point', or 'open a door')
+ */
 public class CommandStackScript {
 	
-	// The data structure of commands
+	// The data structure to hold our commands
 	private Stack<CommandScript> commands;
 	
 	/* Constructor - Creates a new CommandStack */
 	public CommandStackScript() {
 		commands = new Stack<CommandScript>();
+	}
+	
+	/* currentcommand - Stores the command currently at the top of the stack. 
+	 * Returns null if stack is empty */
+	public CommandScript currentCommand {
+		get {
+			if(commands.Count > 0) {
+				return(commands.Peek());
+			}
+			else {
+				// Stack is empty
+				return(null);
+			}
+		}
 	}
 	
 	/* Execute - Run the top command on the stack */
@@ -19,8 +36,7 @@ public class CommandStackScript {
 		if(currentCommand != null) {
 			if(currentCommand.isComplete) {
 				// Current task is finished, pop it off!
-				Debug.Log ("Current command is complete!");
-				commands.Pop();
+				Remove();
 			}
 			else {
 				currentCommand.Execute();
@@ -34,8 +50,6 @@ public class CommandStackScript {
 		if(commands.Count > 0) {
 			// Make sure we're not just receiving the same action again
 			if(currentCommand.GetType() != command.GetType()) {
-				// Pause current command
-				currentCommand.Pause();
 				commands.Push(command);
 			}
 		}
@@ -49,25 +63,7 @@ public class CommandStackScript {
 	public void Remove() {
 		if(currentCommand != null) {
 			commands.Pop ();
-
-			// Check if there is a currentItem
-			if(currentCommand != null) {
-				currentCommand.Resume();
-			}
 		}
 
 	}
-	public CommandScript currentCommand {
-		get {
-			if(commands.Count > 0) {
-				return(commands.Peek());
-			}
-			else {
-				// Stack is empty
-				return(null);
-			}
-		}
-	}
-
-
 }
