+++ Assets/Resources/scripts/characters/ai/commands.meta
@@ -1,0 +1,5 @@
+fileFormatVersion: 2
+guid: 2abb41f0d00344441ada827d02e8421f
+folderAsset: yes
+DefaultImporter:
+  userData:
