+++ Assets/Resources/scripts/characters/ai/commands/GotoCommandScript.cs.meta
